% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuTensor.R
\name{all.gpuTensor}
\alias{all.gpuTensor}
\title{All}
\usage{
\method{all}{gpuTensor}(x, ...)
}
\arguments{
\item{x}{A gpuTensor object (should contain logical values)}

\item{...}{Additional arguments (ignored)}
}
\value{
Logical scalar indicating if all elements are TRUE
}
\description{
Test whether all elements are TRUE in the tensor.
}
