% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuTensor.R
\name{argmin.gpuTensor}
\alias{argmin.gpuTensor}
\title{Argmin for gpuTensor}
\usage{
\method{argmin}{gpuTensor}(x, axis = NULL, keep.dims = FALSE, ...)
}
\arguments{
\item{x}{gpuTensor}

\item{axis}{NULL (global) or integer axis (1-based)}

\item{keep.dims}{logical, keep reduced dims of size 1}
}
\description{
Argmin for gpuTensor
}
