% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuVector.R
\name{as.vector.gpuVector}
\alias{as.vector.gpuVector}
\title{Convert GPU Vector to R Vector}
\usage{
\method{as.vector}{gpuVector}(gpu_vec)
}
\arguments{
\item{gpu_vec}{A gpuVector object}
}
\value{
A numeric vector containing the data from the GPU
}
\description{
Converts a gpuVector object back to an R numeric vector by copying data from GPU to CPU.
}
\details{
This function copies data from GPU memory back to CPU memory and returns
it as a standard R numeric vector.
}
\examples{
\dontrun{
# Create gpuVector and convert back
x <- c(1, 2, 3, 4, 5)
gpu_x <- as.gpuVector(x)
y <- as.vector(gpu_x)
}

}
