% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuTensor.R
\name{as_tensor}
\alias{as_tensor}
\title{Create Tensor from Matrix/Array}
\usage{
as_tensor(x, dtype = "double", requires_grad = FALSE)
}
\arguments{
\item{x}{An R matrix or array}

\item{dtype}{Data type: "double" or "float" (default: "double")}

\item{requires_grad}{Whether to track gradients (default: FALSE)}
}
\value{
A gpuTensor object with the same shape as the input
}
\description{
Creates a gpuTensor from an R matrix or array, preserving dimensions.
}
