% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuTensor.R
\name{concat}
\alias{concat}
\title{Concatenate Tensors}
\usage{
concat(tensors, axis = 1)
}
\arguments{
\item{tensors}{List of gpuTensor objects to concatenate}

\item{axis}{Integer specifying the axis along which to concatenate (1-indexed)}
}
\value{
A new tensor with tensors concatenated along the specified axis
}
\description{
Concatenate tensors along an existing dimension.
}
