% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuTensor.R
\name{cos.gpuTensor}
\alias{cos.gpuTensor}
\title{Cosine Function}
\usage{
\method{cos}{gpuTensor}(x)
}
\arguments{
\item{x}{A gpuTensor object}
}
\value{
A tensor with cos(x) computed element-wise
}
\description{
Computes the cosine of each element in the tensor.
}
