% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuTensor.R
\name{cumprod.gpuTensor}
\alias{cumprod.gpuTensor}
\title{Cumulative Product}
\usage{
\method{cumprod}{gpuTensor}(x)
}
\arguments{
\item{x}{A gpuTensor object}
}
\value{
A gpuTensor with cumulative products
}
\description{
Compute the cumulative product of elements along the tensor.
}
