% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuTensor.R
\name{cumsum.gpuTensor}
\alias{cumsum.gpuTensor}
\title{Cumulative Sum}
\usage{
\method{cumsum}{gpuTensor}(x)
}
\arguments{
\item{x}{A gpuTensor object}
}
\value{
A gpuTensor with cumulative sums
}
\description{
Compute the cumulative sum of elements along the tensor.
}
