% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuTensor.R
\name{diff.gpuTensor}
\alias{diff.gpuTensor}
\title{Differences}
\usage{
\method{diff}{gpuTensor}(x, lag = 1, differences = 1, ...)
}
\arguments{
\item{x}{A gpuTensor object}

\item{lag}{Integer, lag for differences (default 1)}

\item{differences}{Integer, number of differences to compute (default 1)}

\item{...}{Additional arguments (ignored)}
}
\value{
A gpuTensor with differences
}
\description{
Compute differences between consecutive elements.
}
