% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuTensor.R
\name{dtype}
\alias{dtype}
\title{Get Tensor Data Type}
\usage{
dtype(tensor)
}
\arguments{
\item{tensor}{A gpuTensor object}
}
\value{
Character string indicating the data type
}
\description{
Get Tensor Data Type
}
