% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuTensor.R
\name{exp.gpuTensor}
\alias{exp.gpuTensor}
\title{Exponential Function}
\usage{
\method{exp}{gpuTensor}(x)
}
\arguments{
\item{x}{A gpuTensor object}
}
\value{
A tensor with exp(x) computed element-wise
}
\description{
Computes the exponential of each element in the tensor.
}
