% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpu-utils.R
\name{gpu_memory_available}
\alias{gpu_memory_available}
\title{Get Available GPU Memory}
\usage{
gpu_memory_available()
}
\value{
Numeric value of available GPU memory in bytes (0 if no GPU)
}
\description{
Returns the amount of free GPU memory in bytes.
}
\details{
Queries the available GPU memory. Useful for determining if large
operations will fit in GPU memory.
}
\examples{
\dontrun{
mem_gb <- gpu_memory_available() / 1e9
cat("Available GPU memory:", round(mem_gb, 1), "GB\n")
}

}
