% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuTensor.R
\name{>.gpuTensor}
\alias{>.gpuTensor}
\title{Greater Than Comparison}
\usage{
\method{>}{gpuTensor}(a, b)
}
\arguments{
\item{a}{First gpuTensor}

\item{b}{Second gpuTensor}
}
\value{
gpuTensor with numeric 0/1 values
}
\description{
Performs element-wise comparison a > b and returns a tensor of 0/1.
}
