% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuTensor.R
\name{<.gpuTensor}
\alias{<.gpuTensor}
\alias{==.gpuTensor}
\title{Less Than Comparison}
\usage{
\method{<}{gpuTensor}(a, b)

\method{==}{gpuTensor}(a, b)
}
\description{
Less Than Comparison

Equality Comparison
}
