% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuTensor.R
\name{is_contiguous}
\alias{is_contiguous}
\title{Check if Tensor is Contiguous}
\usage{
is_contiguous(tensor)
}
\arguments{
\item{tensor}{A gpuTensor object}
}
\value{
TRUE if tensor has contiguous memory layout
}
\description{
Check if Tensor is Contiguous
}
