% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuVector.R
\name{length.gpuVector}
\alias{length.gpuVector}
\title{Size method for gpuVector objects}
\usage{
\method{length}{gpuVector}(x)
}
\arguments{
\item{x}{A gpuVector object}
}
\value{
The number of elements in the gpuVector
}
\description{
Size method for gpuVector objects
}
