% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuTensor.R
\name{<=.gpuTensor}
\alias{<=.gpuTensor}
\title{Less Than or Equal To Operator for GPU Tensors}
\usage{
\method{<=}{gpuTensor}(x, y)
}
\arguments{
\item{x}{First gpuTensor}

\item{y}{Second gpuTensor}
}
\value{
A gpuTensor with boolean results (1.0 for true, 0.0 for false)
}
\description{
Element-wise less than or equal comparison between two tensors.
Implemented as (x < y) || (x == y).
}
