% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuTensor.R
\name{Math.gpuTensor}
\alias{Math.gpuTensor}
\title{Math Group Generic for GPU Tensors}
\usage{
\method{Math}{gpuTensor}(x, ...)
}
\arguments{
\item{x}{A gpuTensor object}

\item{...}{Additional arguments}
}
\description{
Provides support for mathematical functions on GPU tensors
}
