% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuTensor.R
\name{matmul}
\alias{matmul}
\title{Matrix Multiplication}
\usage{
matmul(a, b)
}
\arguments{
\item{a}{First tensor (2D)}

\item{b}{Second tensor (2D)}
}
\value{
Result of matrix multiplication
}
\description{
Performs matrix multiplication between two 2D tensors.
}
