% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuTensor.R
\name{max.gpuTensor}
\alias{max.gpuTensor}
\title{Maximum Function}
\usage{
\method{max}{gpuTensor}(x, ...)
}
\arguments{
\item{x}{A gpuTensor object}

\item{...}{Additional arguments (ignored)}
}
\value{
The maximum value in the tensor
}
\description{
Finds the maximum element in the tensor.
}
