% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuTensor.R
\name{mean.gpuTensor}
\alias{mean.gpuTensor}
\title{Mean (Average) Function}
\usage{
\method{mean}{gpuTensor}(x, ...)
}
\arguments{
\item{x}{A gpuTensor object}

\item{...}{Additional arguments (ignored)}
}
\value{
The mean value of all tensor elements
}
\description{
Computes the arithmetic mean of all elements in the tensor.
}
