% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuTensor.R
\name{min.gpuTensor}
\alias{min.gpuTensor}
\title{Minimum Function}
\usage{
\method{min}{gpuTensor}(x, ...)
}
\arguments{
\item{x}{A gpuTensor object}

\item{...}{Additional arguments (ignored)}
}
\value{
The minimum value in the tensor
}
\description{
Finds the minimum element in the tensor.
}
