% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuTensor.R
\name{permute}
\alias{permute}
\title{Permute Tensor Dimensions}
\usage{
permute(tensor, dims)
}
\arguments{
\item{tensor}{A gpuTensor object}

\item{dims}{Integer vector specifying the new order of dimensions (1-indexed)}
}
\value{
A tensor with permuted dimensions
}
\description{
Reorder the dimensions of a tensor.
}
