% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuVector.R
\name{+.gpuVector}
\alias{+.gpuVector}
\title{Addition operator for gpuVector objects}
\usage{
\method{+}{gpuVector}(a, b)
}
\arguments{
\item{a}{A gpuVector object}

\item{b}{A gpuVector object}
}
\value{
A gpuVector object containing the sum
}
\description{
Addition operator for gpuVector objects
}
