% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuTensor.R
\name{pmax.gpuTensor}
\alias{pmax.gpuTensor}
\title{Element-wise maximum of two tensors}
\usage{
\method{pmax}{gpuTensor}(a, b, ...)
}
\arguments{
\item{a}{A gpuTensor object}

\item{b}{A gpuTensor object}
}
\value{
A gpuTensor with the element-wise maximum values
}
\description{
Element-wise maximum of two tensors
}
