% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuTensor.R
\name{pmin.gpuTensor}
\alias{pmin.gpuTensor}
\title{Element-wise minimum of two tensors}
\usage{
\method{pmin}{gpuTensor}(a, b, ...)
}
\arguments{
\item{a}{A gpuTensor object}

\item{b}{A gpuTensor object}
}
\value{
A gpuTensor with the element-wise minimum values
}
\description{
Element-wise minimum of two tensors
}
