% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuTensor.R
\name{prod.gpuTensor}
\alias{prod.gpuTensor}
\title{Product (All elements)}
\usage{
\method{prod}{gpuTensor}(x, ...)
}
\arguments{
\item{x}{A gpuTensor object}

\item{...}{Additional args (ignored)}
}
\value{
Numeric scalar
}
\description{
Computes the product of all tensor elements.
}
