% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rand_tensor}
\alias{rand_tensor}
\title{Create Random Tensor (Uniform Distribution)}
\usage{
rand_tensor(shape, dtype = "float")
}
\arguments{
\item{shape}{Integer vector specifying tensor dimensions}

\item{dtype}{String specifying data type ("float" or "double")}
}
\description{
Generate a tensor filled with random numbers from uniform distribution [0, 1).
}
