% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuTensor.R
\name{range.gpuTensor}
\alias{range.gpuTensor}
\title{Range}
\usage{
\method{range}{gpuTensor}(x, ...)
}
\arguments{
\item{x}{A gpuTensor object}

\item{...}{Additional arguments (ignored)}
}
\value{
Numeric vector of length 2 with min and max values
}
\description{
Return the range (min and max) of the tensor values.
}
