% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuTensor.R
\name{relu}
\alias{relu}
\title{ReLU Activation Function}
\usage{
relu(x)
}
\arguments{
\item{x}{A gpuTensor object}
}
\value{
A tensor with relu(x) computed element-wise
}
\description{
Computes the Rectified Linear Unit max(0, x) of each element.
}
