% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuTensor.R
\name{repeat_tensor}
\alias{repeat_tensor}
\title{Repeat Tensor}
\usage{
repeat_tensor(tensor, repeats)
}
\arguments{
\item{tensor}{A gpuTensor object}

\item{repeats}{Integer vector specifying repeat counts for each dimension}
}
\value{
A new tensor with the input tensor repeated
}
\description{
Repeat tensor along specified dimensions.
}
