% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuTensor.R
\name{requires_grad}
\alias{requires_grad}
\title{Enable Gradient Computation}
\usage{
requires_grad(tensor, requires_grad = TRUE)
}
\arguments{
\item{tensor}{A gpuTensor object}

\item{requires_grad}{Boolean indicating whether to track gradients}
}
\value{
The tensor (for method chaining)
}
\description{
Sets whether gradients should be computed for this tensor.
}
