% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuTensor.R
\name{rnorm}
\alias{rnorm}
\title{Random Normal Generic Function}
\usage{
rnorm(x, ...)
}
\arguments{
\item{x}{Object to generate random numbers for}

\item{...}{Additional arguments}
}
\description{
Generate random numbers from normal distribution.
}
