% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuTensor.R
\name{runif}
\alias{runif}
\title{Random Uniform Generic Function}
\usage{
runif(x, ...)
}
\arguments{
\item{x}{Object to generate random numbers for}

\item{...}{Additional arguments}
}
\description{
Generate random numbers from uniform distribution.
}
