% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{set_random_seed}
\alias{set_random_seed}
\title{Set Random Seed}
\usage{
set_random_seed(seed)
}
\arguments{
\item{seed}{Integer seed value}
}
\description{
Set the seed for random number generation.
}
