% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuTensor.R
\name{shape}
\alias{shape}
\title{Get Tensor Shape}
\usage{
shape(tensor)
}
\arguments{
\item{tensor}{A gpuTensor object}
}
\value{
Integer vector containing the tensor dimensions
}
\description{
Get Tensor Shape
}
