% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuTensor.R
\name{sigmoid}
\alias{sigmoid}
\title{Sigmoid Activation Function}
\usage{
sigmoid(x)
}
\arguments{
\item{x}{A gpuTensor object}
}
\value{
A tensor with sigmoid(x) computed element-wise
}
\description{
Computes the sigmoid (logistic) function 1/(1+exp(-x)) of each element.
}
