% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuTensor.R
\name{sin.gpuTensor}
\alias{sin.gpuTensor}
\title{Sine Function}
\usage{
\method{sin}{gpuTensor}(x)
}
\arguments{
\item{x}{A gpuTensor object}
}
\value{
A tensor with sin(x) computed element-wise
}
\description{
Computes the sine of each element in the tensor.
}
