% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuTensor.R
\name{sqrt.gpuTensor}
\alias{sqrt.gpuTensor}
\title{Square Root Function}
\usage{
\method{sqrt}{gpuTensor}(x)
}
\arguments{
\item{x}{A gpuTensor object}
}
\value{
A tensor with sqrt(x) computed element-wise
}
\description{
Computes the square root of each element in the tensor.
}
