% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuTensor.R
\name{stack}
\alias{stack}
\title{Stack Tensors}
\usage{
stack(tensors, axis = 1)
}
\arguments{
\item{tensors}{List of gpuTensor objects to stack}

\item{axis}{Integer specifying the axis along which to stack (1-indexed)}
}
\value{
A new tensor with tensors stacked along the specified axis
}
\description{
Stack tensors along a new dimension.
}
