% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuTensor.R
\name{[<-.gpuTensor}
\alias{[<-.gpuTensor}
\title{Slice assignment for gpuTensor (in-place update on GPU)}
\usage{
\method{[}{gpuTensor}(x, ...) <- value
}
\description{
Currently supports assigning a numeric scalar to a rectangular (contiguous) slice.
More complex assignment (tensor-to-slice) can be added later.
}
