% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuTensor.R
\name{sum.gpuTensor}
\alias{sum.gpuTensor}
\title{Tensor Sum}
\usage{
\method{sum}{gpuTensor}(tensor, ...)
}
\arguments{
\item{tensor}{A gpuTensor object}
}
\value{
Scalar sum of all elements
}
\description{
Computes the sum of all elements in a tensor.
}
