% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuTensor.R
\name{synchronize}
\alias{synchronize}
\title{Synchronize Tensor Operations}
\usage{
synchronize(tensor)
}
\arguments{
\item{tensor}{A gpuTensor object}
}
\description{
Waits for all asynchronous operations on the tensor to complete.
}
