% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuTensor.R
\name{tanh.gpuTensor}
\alias{tanh.gpuTensor}
\title{Hyperbolic Tangent Function}
\usage{
\method{tanh}{gpuTensor}(x)
}
\arguments{
\item{x}{A gpuTensor object}
}
\value{
A tensor with tanh(x) computed element-wise
}
\description{
Computes the hyperbolic tangent of each element in the tensor.
}
