% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuTensor.R
\name{tensor_pow}
\alias{tensor_pow}
\title{Element-wise power of two tensors (a^b)}
\usage{
tensor_pow(a, b)
}
\arguments{
\item{a}{A gpuTensor object (base)}

\item{b}{A gpuTensor object (exponent)}
}
\value{
A gpuTensor with a raised to the power of b element-wise
}
\description{
Element-wise power of two tensors (a^b)
}
