% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuTensor.R
\name{transpose}
\alias{transpose}
\title{Transpose Tensor}
\usage{
transpose(tensor)
}
\arguments{
\item{tensor}{A 2D gpuTensor object}
}
\value{
A transposed tensor that shares memory with the original
}
\description{
Creates a transposed view of a 2D tensor (matrix).
}
