% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuTensor.R
\name{transpose_view}
\alias{transpose_view}
\title{Efficient Transpose View}
\usage{
transpose_view(tensor)
}
\arguments{
\item{tensor}{A 2D gpuTensor object}
}
\value{
A transposed tensor view sharing the same memory
}
\description{
Creates a transpose view that shares memory with the original tensor.
This is much faster than the regular transpose() which copies data.
Similar to PyTorch's .T property.
}
