% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuTensor.R
\name{var.gpuTensor}
\alias{var.gpuTensor}
\title{Variance (Population)}
\usage{
var.gpuTensor(x, ...)
}
\arguments{
\item{x}{A gpuTensor object}

\item{...}{Additional args (ignored)}
}
\value{
Numeric scalar (double)
}
\description{
Computes population variance of tensor elements.
}
