% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuTensor.R
\name{view}
\alias{view}
\title{Create Tensor View}
\usage{
view(tensor, new_shape)
}
\arguments{
\item{tensor}{A gpuTensor object}

\item{new_shape}{Integer vector specifying the new shape}
}
\value{
A new tensor that shares memory with the original
}
\description{
Creates a view of a tensor with a different shape (same data).
}
