% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuTensor.R
\name{which.max.gpuTensor}
\alias{which.max.gpuTensor}
\title{Which Max}
\usage{
which.max.gpuTensor(x, ...)
}
\arguments{
\item{x}{A gpuTensor object}

\item{...}{Additional arguments (ignored)}
}
\value{
Integer index of maximum element (1-based)
}
\description{
Find the index of the maximum element in the tensor.
}
