% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuTensor.R
\name{which.max}
\alias{which.max}
\title{Which Max Generic Function}
\usage{
which.max(x, ...)
}
\arguments{
\item{x}{Object to find maximum index for}

\item{...}{Additional arguments}
}
\description{
Find the index of the maximum element.
}
