% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuTensor.R
\name{which.min.gpuTensor}
\alias{which.min.gpuTensor}
\title{Which Min}
\usage{
which.min.gpuTensor(x, ...)
}
\arguments{
\item{x}{A gpuTensor object}

\item{...}{Additional arguments (ignored)}
}
\value{
Integer index of minimum element (1-based)
}
\description{
Find the index of the minimum element in the tensor.
}
