% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuTensor.R
\name{which.min}
\alias{which.min}
\title{Which Min Generic Function}
\usage{
which.min(x, ...)
}
\arguments{
\item{x}{Object to find minimum index for}

\item{...}{Additional arguments}
}
\description{
Find the index of the minimum element.
}
