% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuTensor.R
\name{as_tensor}
\alias{as_tensor}
\alias{as_tensor.default}
\title{Convert R object to GPU Tensor}
\usage{
as_tensor(x, ...)

\method{as_tensor}{default}(x, dtype = "double", requires_grad = FALSE, shape = NULL, ...)
}
\arguments{
\item{x}{Input R object (vector, matrix, or array)}

\item{...}{Additional arguments passed to methods}

\item{dtype}{Data type for the tensor (default: "double")}

\item{requires_grad}{Whether to track gradients (default: FALSE)}

\item{shape}{Optional shape to reshape the data (default: NULL, uses input shape)}
}
\value{
A gpuTensor object with the same shape as the input
}
\description{
Convert R object to GPU Tensor
}
