% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuTensor.R
\name{[<-.gpuTensor}
\alias{[<-.gpuTensor}
\title{Slice assignment for gpuTensor (in-place update on GPU)}
\usage{
\method{[}{gpuTensor}(x, ...) <- value
}
\description{
Supports assigning scalars or tensors to rectangular slices, and boolean mask assignment.
}
